Rewrite to Typescript
Complete rewrite to typescript.
Modules are loaded with requirejs.
Views extend base React components and are plain es6/TS classes so the
factory method `createClass` is unused.
Instead of MicroEvent I added a small event file because MicroEvent was
not found in DefinitelyTyped and other event emitter libs did not work
or were too complicated when used with requirejs (EventEmitter).